requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto==2.49.0
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
dj-database-url==0.5.0
Django==3.2.3
django-admin-honeypot==1.1.0
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
pyperclip==1.8.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
transfersh-client==1.1.3
urllib3==1.26.8
wget==3.2
whitenoise==5.3.0
